# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lee Pump App API swagger documentation",
        "description": "This is a sample API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/upload/upload-file": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "upload file function ",
                "description": "upload file function",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "description": "file",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/changePassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "description": "Change user password",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "name": "current_password",
                        "in": "query",
                        "description": "current password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "new password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully changed password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/global/country_list": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "list of country",
                "description": "list of country",
                "operationId": "country_list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false
            }
        },
...
        "/api/member/get-member-info": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "",
                "description": "member info",
                "operationId": "getMemberInfo",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/change-password": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "change password ",
                "description": "change password",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "old_password",
                        "in": "query",
                        "description": "old_password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password_confirmation",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/set-secpassword": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "set sec password ",
                "description": "set sec password",
                "operationId": "setSecPassword",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password_confirmation",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/change-secpassword": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "change sec password ",
                "description": "change sec password",
                "operationId": "changeSecPassword",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sec_password",
                        "in": "query",
                        "description": "sec_password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password_confirmation",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/member/user-bank": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "set user bank ",
                "description": "set user bank",
                "operationId": "userBank",
                "parameters": [
                    {
                        "name": "sec_password",
                        "in": "query",
                        "description": "sec_password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bank_country",
                        "in": "query",
                        "description": "country id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bank_name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bank_user",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bank_number",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "swift_code",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/member/get-bank-info": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "get user bank ",
                "description": "get user bank",
                "operationId": "getBankInfo",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/package/get-package": {
            "get": {
                "tags": [
                    "Package"
                ],
                "summary": "",
                "operationId": "getPackage",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}